+++ src/Network/Network/ChangeLog.md
@@ -38,6 +38,10 @@
 	- `Get-AzVpnClientConfiguration`
 	- `New-AzVirtualNetworkGateway`
 	- `Set-AzVirtualNetworkGateway`
+* Updated cmdlets to enable setting of PrivateRange on AzureFirewallPolicy.
+    - `New-AzFirewallPolicy`
+    - `Set-AzFirewallPolicy`
+
 
 ## Version 4.6.0
 * Added new cmdlets to replace old product name `virtual router` with new name `route server` in the future.